requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
pinecone-client==2.2.4
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.43.0
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
joblib==1.3.2
imbalanced-learn==0.11.0
beautifulsoup4==4.12.2
newsapi-python==0.2.7
reportlab==4.0.7
fpdf2==2.7.6
pydantic-settings==2.1.0
apscheduler==3.10.4
python-multipart==0.0.6
tiktoken==0.5.2
aiosmtplib==3.0.1
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.1.0
